chore(TechDocs): extract entities to docs mapper to be shared amount DocsTable(s)
